backend/requirements.txt
attrs==24.2.0
blinker==1.8.2
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.0.3
reportlab==4.0.8
python-dateutil>=2.8.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
mistune==3.0.2
mpmath==1.3.0
packaging==24.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
six==1.16.0
Werkzeug==3.0.4